+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 
-## 0.15.0 (tbd)
+## 0.15.1 (tbd)
+
+- Fixed issue with `piral-core` 0.15 onwards (Webpack v4 does not respect `exports` section of *package.json*)
+- Updated `pilet-webpack-plugin` to allow placeholders in entry module file name 
+
+## 0.15.0 (November 17, 2022)
 
 - Moved `piral-cli-webpack` package to dedicated repository
-- Updated `pilet-webpack-plugin` to allow placeholders in entry module file name 
+- Updated Lerna to v6